requirements.txt
Flask==1.1.2
gunicorn==19.6.0
requests==2.23.0
Flask-SQLAlchemy==2.4.3
python-chess==0.30.1
SQLAlchemy==1.3.17